#27 rename InfraWalletApi to InfraWalletClient
